Adding a new class to Drag elements
This new class will substitute the usage of a HOC. I have decided to create a new class that receives only the id of the dialog, so this class only makes the dialog draggable, he doesnt worry about create a container or pass CSS properties to dialog. This turn the usage more simple to understand and remove the dependence of the react.